Add route for genre selection, changed className for genre-card to avoid error
