+++ frontend/src/components/directory/MovieCardList.jsx
@@ -19,14 +19,6 @@ const MovieCardList = ({
 }) => {
   const navigate = useNavigate();
 
-  // Helper function to check if value exists and is not empty
-  const hasValue = (value) => {
-    if (value === undefined || value === null) return false;
-    if (typeof value === 'string') return value.trim() !== '';
-    if (Array.isArray(value)) return value.length > 0 && value[0] !== '';
-    return true;
-  };
-
   const handleCardClick = () => {
     navigate(`/movie/${encodeURIComponent(movie.title)}`, {
       state: { movieData: movie }, // Pass entire movie object
@@ -75,21 +67,19 @@ const MovieCardList = ({
         />
       </div>
       <div className="movie-details-container-list">
-        <h3 style={{paddingInlineStart:7}}>{movie.title}</h3>
+        <h3>{movie.title}</h3>
         <ReviewStars
           rating={movie.rating}
           readOnly={true}
           showNumber={showRatingNumber}
         />
-        {hasValue(movie.genre) && (
-          <div className="genre-tags">
-            {movie.genre.map((genre, index) => (
-              <span key={index} className="genre-tag">
-                {genre}
-              </span>
-            ))}
-          </div>
-        )}
+        <div className="genre-tags">
+          {movie.genre.map((genre, index) => (
+            <span key={index} className="genre-tag">
+              {genre}
+            </span>
+          ))}
+        </div>
         <div className="duration-and-icons-container">
           <div className="duration-tag">
             <span className="duration-icon">
@@ -114,33 +104,25 @@ const MovieCardList = ({
             </div>
           )}
         </div>
-        {showCastInfo && (hasValue(movie.director) || hasValue(movie.actors)) && (
+        {showCastInfo && (
           <div className="cast-info">
-            {hasValue(movie.director) && (
-              <div className="cast-row">
-                <span className="cast-label">Director</span>
-                <span className="director-item">{movie.director}</span>
+            <div className="cast-row">
+              <span className="cast-label">Director</span>
+              <span className="director-item">{movie.director}</span>
+            </div>
+            <div className="cast-row">
+              <span className="cast-label">Cast</span>
+              <div className="actors-list">
+                {movie.actors.map((actor, index) => (
+                  <span key={index} className="actor-item">
+                    {actor}
+                  </span>
+                ))}
               </div>
-            )}
-
-            {hasValue(movie.actors) && (
-              <div className="cast-row">
-                <span className="cast-label">Cast</span>
-                <div className="actors-list">
-                  {movie.actors.map((actor, index) => (
-                    <span key={index} className="actor-item">
-                      {actor}
-                    </span>
-                  ))}
-                </div>
-              </div>
-            )}
+            </div>
           </div>
         )}
-
-        {hasValue(movie.description) && (
-          <p className="clamp-text">{movie.description}</p>
-        )}
+        <p className="clamp-text">{movie.description}</p>
         {showBottomInteractiveIcon && (
           <div
             className="bottom-iteractive-icon-container"
@@ -152,7 +134,6 @@ const MovieCardList = ({
             </button>
             <div className="iteractive-icon" onClick={handleLikeClick}>
               <LikeIcon liked={liked} />
-              <span className="like-count">{movie.likes}</span>
             </div>
             <div
               className="iteractive-icon"
